Split up render-string-template from getFieldsFromTemplate
