+++ app/src/interfaces/tree-view/tree-view.vue
@@ -50,7 +50,7 @@ import { defineComponent, ref, computed, PropType, onMounted, watch } from '@vue
 import { useCollection } from '@/composables/use-collection';
 import { useRelationsStore } from '@/stores';
 import api from '@/api';
-import { getFieldsFromTemplate } from '@/utils/render-template';
+import { getFieldsFromTemplate } from '@/utils/get-fields-from-template';
 import draggable from 'vuedraggable';
 import hideDragImage from '@/utils/hide-drag-image';
 import NestedDraggable from './nested-draggable.vue';
