+++ app/src/utils/get-fields-from-template.ts
@@ -1,0 +1,15 @@
+export function getFieldsFromTemplate(template: string | null) {
+	if (template === null) return [];
+
+	const regex = /{{(.*?)}}/g;
+	let fields = template.match(regex);
+
+	if (!Array.isArray(fields)) {
+		return [];
+	}
+
+	fields = fields.map((field) => {
+		return field.replace(/{{/g, '').replace(/}}/g, '').trim();
+	});
+	return fields as string[];
+}
