--- app/src/utils/render-template/index.ts
@@ -1,4 +1,0 @@
-import { getFieldsFromTemplate, renderTemplate } from './render-template';
-
-export { getFieldsFromTemplate, renderTemplate };
-export default renderTemplate;
